core: Move TextRenderSettings to font/text_render_settings.rs
This makes font.rs smaller, as it's a pretty big file.
